+++ DiffEqDLL/main.cpp
@@ -2,5 +2,5 @@
 void main()
 {
 	myClass sumTwoClass(10, 5);
-	double result = sumTwoClass.sumX_Y();
+	int result = sumTwoClass.sumX_Y();
 }
